% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircoTax.R
\name{CircoTax}
\alias{CircoTax}
\title{CircoTax Plot}
\usage{
CircoTax(
  input_table,
  title = "CircoTax plot",
  fill_text = "log2FC",
  ramp = c("blue", "white", "orange"),
  tax_col = 2:length(colnames(input_table)),
  fc_col = ifelse(length(tax_col) > 1, 1, 3),
  names = 1,
  size_taxon_circo = 4,
  sort = c("rank", "no", "fc", "absfc", "alpha", "alpharank"),
  rank_vector = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  decrease_order = FALSE,
  collapse_ranks = FALSE,
  save_path = NULL,
  image_dim = 4000,
  save_format = "png"
)
}
\arguments{
\item{input_table}{Input DataFrame}

\item{title}{Title display on the plot. Default: CircoTax plot}

\item{fill_text}{Text to display above the color legend. Default: log2FC}

\item{ramp}{Character vector to represent a color gradient for the log2FC scale of values, from min to max. Default: Defaults to c("blue","white","orange")}

\item{tax_col}{Column index or range of column indexes holding IDs for the ranks. Default: 2:length(colnames(input_table))}

\item{fc_col}{Column index holding log2FC values. Default: 1 when tax_col argument is a range of indexes else 3}

\item{names}{Column index holding IDs for the taxon, used when tax_col argument is a single index. Default: 1}

\item{size_taxon_circo}{Size of plot annotation. Default: 4}

\item{sort}{Sorting logic: "no" (no sorting), "rank" (by taxonomic rank), "fc" (by fold change), "absfc" (by absolute fold change), "alpha" (alphabetic order), "alpharank" (). Default: "rank".}

\item{rank_vector}{Character vector holding the ranking system ordered from Highest rank to lowest. Default: c("Domain","Phylum","Class","Order","Family","Genus","Species")}

\item{decrease_order}{Boolean to invert order in sorting. Default: FALSE}

\item{collapse_ranks}{Boolean to use only Ranks represented in the table. Default: TRUE}

\item{save_path}{Save path for plot image. Default: NULL}

\item{image_dim}{Pixel value for height and width of the plot.  Default: 4000}

\item{save_format}{"png" or "pdf". Default: "png"}
}
\value{
CircoTax plot
}
\description{
CircoTax is an R function implementing a specialized ggplot2 graph to represent in a radial form
a rank-aware collection of differentially abundant taxa.
}
