% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircoTaxAutoDA.R
\name{CircoTaxAutoDA}
\alias{CircoTaxAutoDA}
\title{CircoTax AutoDA}
\usage{
CircoTaxAutoDA(
  data,
  contrast,
  design = NULL,
  tax_out = NULL,
  mode = c("DESeq2", "ALDEx2"),
  MCS = 128,
  p_adjustment = "BH",
  eff_size = 0,
  p_val = 0.05,
  B = 50,
  lfc = 1,
  lfc_shrink_method = "none",
  remove_redundants = TRUE,
  remove_results = ""
)
}
\arguments{
\item{data}{phyloseq input object}

\item{contrast}{Vector of three characters which are (in this order) the names of the factor of interest and two of its levels of which report the differences. e.g. c("SEX","M","F")}

\item{design}{String describing statistical design (within the limits of DESeq2 or ALDEx2) on which the analysis is conducted. e.g. '~ Gender+Condition}

\item{tax_out}{Vector whose characters are the taxonomic level to remove. Default: c()}

\item{mode}{Differential Analysis method to be used, either "DESeq2" or "ALDEx2". Default: "DESeq2"}

\item{MCS}{Monte-Carlo samples to generate during ALDEx2 computations. Default: 128.}

\item{p_adjustment}{Multiple test adjustment by adjusting (penalize) each p-value. The possible inputs are "BH" (Benjamini-Hochberg) and "holm" (Holm). Default: "BH".}

\item{eff_size}{ALDEx2 effect size threshold under which exclude results. Default: 0.}

\item{p_val}{p-value threshold (after the multiple test adjustment) to consider a result as significant. Default: 0.05}

\item{B}{DESeq2 base-mean threshold (mean after DESeq2 values transformation) under which exclude results. Default: 50.}

\item{lfc}{Log fold change threshold. Default: 1}

\item{lfc_shrink_method}{Shrinkage method of the log fold change, either "none", "apeglm" and "ashr". Default: "none".}

\item{remove_redundants}{Boolean to enable the automatic removal of redundant results (e.g. result repeated also in higher taxonomic levels being the only observation in that taxonomic clade). Default: TRUE.}

\item{remove_results}{Vector whose characters are the bacteria name to remove. Default: c()}
}
\value{
Dataframe
}
\description{
Starting from a phyloseq object, performs the differential analysis at each taxonomic level between two groups using either DESeq2 or ALDEx2.
They then filter the significant results and generate a tsv table, ready to be displayed with CircoTax().
}
